# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentation de l'API hackapi"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Serveur local"
    }
  ],
  "paths": {
...
    "/features/secure-password-generator": {
      "get": {
        "summary": "Génère un mot de passe sécurisé",
        "tags": [
          "Feature - Secure password generator"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "length",
            "required": false,
            "description": "La longueur du mot de passe à générer. Par défaut, il est de 12 caractères.",
            "schema": {
              "type": "integer",
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mot de passe généré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "password": {
                      "type": "string",
                      "description": "Le mot de passe généré"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Longueur du mot de passe invalide",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Message d'erreur concernant la longueur"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Message d'erreur pour tout problème survenu"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}